ProfileIndexationGenerator to IndexationGenerator1D
