+++ pyxem/generators/indexation_generator1d.py
@@ -1,0 +1,70 @@
+# -*- coding: utf-8 -*-
+# Copyright 2017-2019 The pyXem developers
+#
+# This file is part of pyXem.
+#
+# pyXem is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# pyXem is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with pyXem.  If not, see <http://www.gnu.org/licenses/>.
+
+"""Indexation generator and associated tools.
+
+"""
+
+import numpy as np
+
+from pyxem.utils.indexation_utils import index_magnitudes
+
+
+class IndexationGenerator1D():
+    """Generates an indexer for data using a number of methods.
+
+    Parameters
+    ----------
+    profile : ElectronDiffraction1D
+        The signal of diffraction profiles to be indexed.
+    library : ProfileSimulation
+        The simulated profile data.
+
+    """
+
+    def __init__(self, magnitudes, simulation, mapping=True):
+        self.map = mapping
+        self.magnitudes = magnitudes
+        self.simulation = simulation
+
+    def index_peaks(self,
+                    tolerance=0.1,
+                    *args,
+                    **kwargs):
+        """Assigns hkl indices to peaks in the diffraction profile.
+
+        Parameters
+        ----------
+        tolerance : float
+            The n orientations with the highest correlation values are returned.
+        keys : list
+            If more than one phase present in library it is recommended that
+            these are submitted. This allows a mapping from the number to the
+            phase.  For example, keys = ['si','ga'] will have an output with 0
+            for 'si' and 1 for 'ga'.
+        *args : arguments
+            Arguments passed to the map() function.
+        **kwargs : arguments
+            Keyword arguments passed to the map() function.
+
+        Returns
+        -------
+        matching_results : ProfileIndexation
+
+        """
+        return index_magnitudes(np.array(self.magnitudes), self.simulation, tolerance)
